// js/stat-evaluation-player/src/generated/FiftyFiftyEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type FiftyFiftyEvent = { start_time: number, start_frame: number, resolve_time: number, resolve_frame: number, is_kickoff: boolean, team_zero_player: RemoteIdTs | null, team_one_player: RemoteIdTs | null, team_zero_position: [number, number, number], team_one_position: [number, number, number], midpoint: [number, number, number], plane_normal: [number, number, number], winning_team_is_team_0: boolean | null, possession_team_is_team_0: boolean | null, };