// js/stat-evaluation-player/src/generated/FiftyFiftyTeamStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Per-team accumulated 50/50 stats.
 */
export type FiftyFiftyTeamStats = { count: number, wins: number, losses: number, neutral_outcomes: number, kickoff_count: number, kickoff_wins: number, kickoff_losses: number, kickoff_neutral_outcomes: number, possession_after_count: number, opponent_possession_after_count: number, neutral_possession_after_count: number, kickoff_possession_after_count: number, kickoff_opponent_possession_after_count: number, kickoff_neutral_possession_after_count: number, };